Add same bus/port "affinity" to prevent timeouts when booting multiple RPis
If multiple RPis are booting in USB mode at the same time, rpiboot now
requires that one Pi at a time is serviced from sending bootcode.bin
through to the end of the file server phase.  This insures that there
is no chance for bootcode.bin to time out if bootcode.bin is sent and
other RPis serviced before that first Pi has a chance to complete fileserver
mode.